+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Due to this more methods are now synchronous
 * Add more properties to `PackageInfo` and use `DateTime` when applicable
 * Add `getIcon` and `getIconFromApk` method
+* Add `getInstalledApps` method to retrieve all installed apps
 * The method `getPackageNameFromApk` is now called `getAppInfoFromApk` and returns `PackageInfo`
 
 ## 0.0.1